+++ vertx-core/src/main/java/io/vertx/core/http/impl/HttpClientPush.java
@@ -48,7 +48,7 @@ public class HttpClientPush implements HttpRequest {
   }
 
   @Override
-  public int id() {
+  public long id() {
     return stream.id();
   }
